requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.26.2

# Time Series Forecasting
prophet==1.1.5
statsmodels==0.14.1

# Data Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Dashboard Framework
streamlit==1.29.0

# Utilities
python-dateutil==2.8.2
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Optional: SARIMA comparison (if you want to implement it later)
pmdarima==2.0.4
